port: the matrix builders call sinf/cosf, not double sin/cos rounded
PSMTXRotAxisRad and MTXRotRad went through double sin/cos and rounded the
result. The SDK calls the single-precision sinf/cosf, which since ab54f3fd9
are MSL's polynomial here -- a different function, landing a bit away.

Neutral on the measurement: still 2 differing float columns, 3 frames out of
104, both ecbby, earliest frame 71. The fighter joints do not reach their
matrices through these, so this closes nothing by itself; it is the same
argument as the trig change, that the port should compute what the console
computes.

+++ src/pc_stub/psmath_pc.c
@@ -162,8 +162,11 @@ void PSMTXRotAxisRad(Mtx m, Vec* axis, f32 rad)
     if (!ps_ptr_valid(m) || !ps_ptr_valid(axis)) return;
     Vec n;
     PSVECNormalize(axis, &n);
-    f32 s = (f32)sin((double)rad);
-    f32 c = (f32)cos((double)rad);
+    /* The SDK calls the single-precision sinf/cosf, which are MSL's
+     * polynomial. Going through double sin/cos and rounding at the end is a
+     * different function and lands a bit away. */
+    f32 s = sinf(rad);
+    f32 c = cosf(rad);
     f32 t = 1.0f - c;
     f32 x = n.x, y = n.y, z = n.z;
     f32 xSq = x * x, ySq = y * y, zSq = z * z;
@@ -333,5 +336,6 @@ void MTXRotRad(Mtx m, char axis, f32 rad)
     if (!ps_ptr_valid(m)) {
         return;
     }
-    MTXRotTrig(m, axis, (f32) sin((double) rad), (f32) cos((double) rad));
+    /* sinf/cosf, not double sin/cos rounded -- see PSMTXRotAxisRad. */
+    MTXRotTrig(m, axis, sinf(rad), cosf(rad));
 }
